+++ api/Reforge.Infra/ReforgeDbContext.cs
@@ -1,5 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 using Reforge.Core.Activities.Domain;
+using Reforge.Core.ClosedDays.Domain;
 using Reforge.Core.MealLibrary.Domain;
 using Reforge.Core.Meals.Domain;
 using Reforge.Core.Measurements.Domain;
@@ -21,6 +22,7 @@ public class ReforgeDbContext(DbContextOptions<ReforgeDbContext> options) : DbCo
     public DbSet<Measurement> Measurements => Set<Measurement>();
     public DbSet<ReminderSettings> ReminderSettings => Set<ReminderSettings>();
     public DbSet<CustomReminder> CustomReminders => Set<CustomReminder>();
+    public DbSet<ClosedDay> ClosedDays => Set<ClosedDay>();
 
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {