+++ api/Reforge.Infra/ServiceRegistration.cs
@@ -2,6 +2,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Reforge.Core.Activities.OutputPorts;
+using Reforge.Core.ClosedDays.OutputPorts;
 using Reforge.Core.MealLibrary.OutputPorts;
 using Reforge.Core.Meals.OutputPorts;
 using Reforge.Core.Measurements.OutputPorts;
@@ -30,6 +31,7 @@ public static class ServiceRegistration
         services.AddScoped<IMeasurementRepository, MeasurementRepository>();
         services.AddScoped<IReminderSettingsRepository, ReminderSettingsRepository>();
         services.AddScoped<ICustomReminderRepository, CustomReminderRepository>();
+        services.AddScoped<IClosedDayRepository, ClosedDayRepository>();
 
         services.AddScoped<IClock, SystemClock>();
         services.AddScoped<IIdGenerator, SystemGuidIdGenerator>();